GitHub Actions workflow to build HTML documentation and deploy to GitHub Pages (#3)
The workflow's build job installs dependencies in a Conda environment, builds the HTML documentation using sphinx-build and uploads an artifact containing the built documentation. This is triggered by the push event.

The workflow's deploy job downloads the built documentation artifact, makes some minor changes, and deploys to GitHub Pages using JamesIves/github-pages-deploy-action. This is triggered by the push event, but restricted to only run for ACRC:main.

The permissions for GITHUB_TOKEN have been restricted to the minimum required for each job.